+++ NEWS.md
@@ -1,7 +1,6 @@
 # downlit (development version)
 
-* New R pipe `|>` is now highlighted correctly as an infix operator like
-  magrittr's `%>%` (#126).
+* Adds support for new R pipe `|>` syntax (#126).
 
 # downlit 0.4.0