+++ NEWS.md
@@ -1,4 +1,4 @@
-# downlit (development version)
+# downlit 0.4.1
 
 ## Syntax highlighting